Fix error on Windows builds (conversion from 'unsigned long' to 'uint8_t')

+++ programs/x509/cert_req.c
@@ -117,7 +117,7 @@ struct options {
 static void ip_string_to_bytes(const char *str, uint8_t *bytes, int maxBytes)
 {
     for (int i = 0; i < maxBytes; i++) {
-        bytes[i] = strtoul(str, NULL, 16);
+        bytes[i] = (uint8_t) strtoul(str, NULL, 16);
         str = strchr(str, '.');
         if (str == NULL || *str == '\0') {
             break;
